% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_xaringan_media_calls.R
\name{ds_xaringan_media_calls}
\alias{ds_xaringan_media_calls}
\title{Create List of All Media Calls in a Xaringan Slide Deck}
\usage{
ds_xaringan_media_calls(rmd_file)
}
\arguments{
\item{rmd_file}{Path and filename of the Rmarkdown file to be processed.  Rmd file must be present in current working directory}
}
\value{
List of all media calls in markdown file printed to console
}
\description{
Create List of All Media Calls in a Xaringan Slide Deck
}
